Move taproot hash types to primitives
Move the three `taproot` hash types and their associated tags over to
the `primitives` crate.
